reverted whitespace

+++ nussl/datasets/hooks.py
@@ -22,18 +22,18 @@ from ..core import constants, utils
 from .base_dataset import BaseDataset, DataSetException
 
 
+
+
 class MUSDB18(BaseDataset):
     """
     Hook for MUSDB18. Uses the musdb.DB object to access the
     dataset. If ``download=True``, then the 7s snippets of each track
     are downloaded to ``self.folder``. If no folder is given, then
-    the tracks are downloaded to ~/.nussl/musdb18.
-
-    Getting an item from this dataset with no transforms returns the
+    the tracks are downloaded to ~/.nussl/musdb18. 
+    
+    Getting an item from this dataset with no transforms returns the 
     following dictionary:
-
     .. code-block:: none
-
         {
             'mix': [AudioSignal object containing mix audio],
             'source': {
@@ -46,27 +46,27 @@ class MUSDB18(BaseDataset):
                 'labels': ['bass', 'drums', 'other', 'vocals']
             }
         }
-
+    
     Args:
-        folder (str, optional): Location that should be processed to produce the
+        folder (str, optional): Location that should be processed to produce the 
             list of files. Defaults to None.
-        is_wav (bool, optional):  Expect subfolder with wav files for each source
+        is_wav (bool, optional):  Expect subfolder with wav files for each source 
             instead of stems, defaults to False.
-        download (bool, optional): Download sample version of MUSDB18 which
+        download (bool, optional): Download sample version of MUSDB18 which 
             includes 7s excerpts. Defaults to False.
-        subsets (list, optional): Select a musdb subset train or test.
+        subsets (list, optional): Select a musdb subset train or test. 
             Defaults to ['train', 'test'] (all tracks).
-        split (str, optional): When subset train is loaded, split selects the
-            train/validation split. split=’train’ loads the training split,
-            `split=’valid’ loads the validation split. split=None applies no
+        split (str, optional): When subset train is loaded, split selects the 
+            train/validation split. split=’train’ loads the training split, 
+            `split=’valid’ loads the validation split. split=None applies no 
             splitting. Defaults to None.
-        **kwargs: Any additional arguments that are passed up to BaseDataset
+        **kwargs: Any additional arguments that are passed up to BaseDataset 
             (see ``nussl.datasets.BaseDataset``).
     """
     DATASET_HASHES = {
         "musdb": "56777516ad56fe6a8590badf877e6be013ff932c010e0fbdb0aba03ef878d4cd",
     }
-
+    
     def __init__(self, folder=None, is_wav=False, download=False,
                  subsets=None, split=None, **kwargs):
         subsets = ['train', 'test'] if subsets is None else subsets
@@ -74,7 +74,7 @@ class MUSDB18(BaseDataset):
             folder = os.path.join(
                 constants.DEFAULT_DOWNLOAD_DIRECTORY, 'musdb18'
             )
-        self.musdb = musdb.DB(root=folder, is_wav=is_wav, download=download,
+        self.musdb = musdb.DB(root=folder, is_wav=is_wav, download=download, 
                               subsets=subsets, split=split)
         super().__init__(folder, **kwargs)
 
@@ -88,7 +88,7 @@ class MUSDB18(BaseDataset):
         self._setup_audio_signal(mix)
         for source in list(sources.values()):
             self._setup_audio_signal(source)
-
+        
         output = {
             'mix': mix,
             'sources': sources,
@@ -102,9 +102,7 @@ class MUSDB18(BaseDataset):
 class MixSourceFolder(BaseDataset):
     """
     This dataset expects your data to be formatted in the following way:
-
     .. code-block:: none
-
         data/
             mix/
                 [file0].wav
@@ -127,19 +125,15 @@ class MixSourceFolder(BaseDataset):
                 [file2].wav
                 ...
             ...
-
     Note that the the filenames match between the mix folder and each source folder.
-    The source folder names can be whatever you want. Given a file in the
-    ``self.mix_folder`` folder, this dataset will look up the corresponding files
-    with the same name in the source folders. These are the source audio files.
-    The sum of the sources should equal the mixture. Each source will be labeled
+    The source folder names can be whatever you want. Given a file in the 
+    ``self.mix_folder`` folder, this dataset will look up the corresponding files 
+    with the same name in the source folders. These are the source audio files. 
+    The sum of the sources should equal the mixture. Each source will be labeled 
     according to the folder name it comes from.
-
-    Getting an item from this dataset with no transforms returns the
+    Getting an item from this dataset with no transforms returns the 
     following dictionary:
-
     .. code-block:: none
-
         {
             'mix': [AudioSignal object containing mix audio],
             'source': {
@@ -153,18 +147,16 @@ class MixSourceFolder(BaseDataset):
                 'labels': ['label0', 'label1', 'label2', 'label3']
             }
         }
-
-
     Args:
-        folder (str, optional): Location that should be processed to produce the
+        folder (str, optional): Location that should be processed to produce the 
             list of files. Defaults to None.
         mix_folder (str, optional): Folder to look in for mixtures. Defaults to 'mix'.
-        source_folders (list, optional): List of folders to look in for sources.
-            Path is defined relative to folder. If None, all folders other than
+        source_folders (list, optional): List of folders to look in for sources. 
+            Path is defined relative to folder. If None, all folders other than 
             mix_folder are treated as the source folders. Defaults to None.
-        ext (list, optional): Audio extensions to look for in mix_folder.
+        ext (list, optional): Audio extensions to look for in mix_folder. 
             Defaults to ['.wav', '.flac', '.mp3'].
-        **kwargs: Any additional arguments that are passed up to BaseDataset
+        **kwargs: Any additional arguments that are passed up to BaseDataset 
             (see ``nussl.datasets.BaseDataset``).
     """
     def __init__(self, folder, mix_folder='mix', source_folders=None, sample_rate=None,
@@ -199,7 +191,7 @@ class MixSourceFolder(BaseDataset):
             source_path = os.path.join(self.folder, k, item)
             if os.path.exists(source_path):
                 sources[k] = self._load_audio_file(source_path)
-
+        
         if self.make_mix:
             mix = sum(list(sources.values()))
         else:
@@ -226,38 +218,28 @@ class Scaper(BaseDataset):
     files which specify the parameters of the soundscape. If the soundscape is
     generated with `save_isolated_events=True`, then the audio corresponding
     to each event in the soundscape will be saved as well.
-
-    Below is an example of using Scaper to generate a small dataset of 10
+    Below is an example of using Scaper to generate a small dataset of 10 
     mixtures with 2 sources each. The generated dataset can then be immediately
     loaded into an instance of ``nussl.datasets.Scaper`` for integration into
     a training or evaluation pipeline.
-
     The sources are output in a dictionary that looks like this:
-
     .. code-block:: none
-
         data['sources] = {
             '{label}::{count}': AudioSignal,
             '{label}::{count}': AudioSignal,
             ...
         }
-
     For example:
-
     .. code-block:: none
-
         data['sources] = {
             'siren::0': AudioSignal,
             'siren::1': AudioSignal,
             'car_horn::0': AudioSignal,
             ...
         }
-
-    Getting an item from this dataset with no transforms returns the
+    Getting an item from this dataset with no transforms returns the 
     following dictionary:
-
     .. code-block:: none
-
         {
             'mix': [AudioSignal object containing mix audio],
             'source': {
@@ -272,10 +254,7 @@ class Scaper(BaseDataset):
                 'labels': ['label0', 'label1', 'label2', 'label3']
             }
         }
-
-
     Example of generating a Scaper dataset and then loading it with nussl:
-
     >>> n_sources = 2
     >>> n_mixtures = 10
     >>> duration = 3
@@ -303,7 +282,7 @@ class Scaper(BaseDataset):
     >>>     sc.generate(audio_path, jams_path, save_isolated_events=True)
     >>> dataset = nussl.datasets.Scaper(output_dir)
     >>> dataset[0] # contains mix, sources, and metadata corresponding to 0.jams.
-
+        
     Raises:
         DataSetException: if Scaper dataset wasn't saved with isolated event audio.
     """
@@ -354,15 +333,15 @@ class OnTheFly(BaseDataset):
     """
     Hook for a dataset that creates mixtures on the fly from source
     data. The function that creates the mixture is a closure which
-    is defined by the end-user. The number of mixtures in the
+    is defined by the end-user. The number of mixtures in the 
     dataset is also defined by the end-user. The mix closure function
-    should take two arguments - the dataset object and the index of the
-    item being processed - and the output of the mix closure should be a
-    dictionary containing at least a 'mix', 'sources' and (optionally)
-    a 'metadata' key, or other keys that can be defined up to you.
-    Here's an example of a closure, which can be configured via
+    should take two arguments - the dataset object and the index of the 
+    item being processed - and the output of the mix closure should be a 
+    dictionary containing at least a 'mix', 'sources' and (optionally) 
+    a 'metadata' key, or other keys that can be defined up to you. 
+    Here's an example of a closure, which can be configured via 
     variable scoping:
-
+    
     >>>  def make_sine_wave(freq, sample_rate, duration):
     >>>      dt = 1 / sample_rate
     >>>      x = np.arange(0.0, duration, dt)
@@ -387,19 +366,18 @@ class OnTheFly(BaseDataset):
     >>>          'mix': mix,
     >>>          'sources': sources,
     >>>          'metadata': {
-    >>>              'frequencies': freqs
-    >>>          }
+    >>>              'frequencies': freqs    
+    >>>          }    
     >>>      }
     >>>      return output
     >>>  dataset = nussl.datasets.OnTheFly(make_mix, 10)
-
     Args:
         mix_closure (function): A closure that determines how to create
-          a single mixture, given the index. It has a strict input
+          a single mixture, given the index. It has a strict input 
           signature (the index is given as an int) and a strict output
           signature (a dictionary containing a 'mix' and 'sources') key.
         num_mixtures (int): Number of mixtures that will be created on
-          the fly. This determines one 'run' thrugh the dataset, or an
+          the fly. This determines one 'run' thrugh the dataset, or an 
           epoch.
         kwargs: Keyword arguments to BaseDataset.
     """
@@ -411,7 +389,7 @@ class OnTheFly(BaseDataset):
 
     def get_items(self, folder):
         return list(range(self.num_mixtures))
-
+    
     def process_item(self, item):
         output = self.mix_closure(self, item)
         if not isinstance(output, dict):
@@ -424,33 +402,28 @@ class OnTheFly(BaseDataset):
 
 class FUSS(Scaper):
     """
-    The Free Universal Sound Separation (FUSS) Dataset is a database of arbitrary
-    sound mixtures and source-level references, for use in experiments on
-    arbitrary sound separation.
-
-    This is the official sound separation data for the DCASE2020 Challenge Task 4:
+    The Free Universal Sound Separation (FUSS) Dataset is a database of arbitrary 
+    sound mixtures and source-level references, for use in experiments on 
+    arbitrary sound separation. 
+    This is the official sound separation data for the DCASE2020 Challenge Task 4: 
     Sound Event Detection and Separation in Domestic Environments.
-
-    This is a hook for reading in this dataset, and making sure that the mix and
+    This is a hook for reading in this dataset, and making sure that the mix and 
     source paths are massaged to be relative paths.
-
     References:
-
-    [1]  Scott Wisdom, Hakan Erdogan, Daniel P. W. Ellis, Romain Serizel,
-    Nicolas Turpault, Eduardo Fonseca, Justin Salamon, Prem Seetharaman,
-    John R. Hershey, "What's All the FUSS About Free Universal Sound Separation
+    [1]  Scott Wisdom, Hakan Erdogan, Daniel P. W. Ellis, Romain Serizel, 
+    Nicolas Turpault, Eduardo Fonseca, Justin Salamon, Prem Seetharaman, 
+    John R. Hershey, "What's All the FUSS About Free Universal Sound Separation 
     Data?", 2020, in preparation.
-
-    [2] Eduardo Fonseca, Jordi Pons, Xavier Favory, Frederic Font Corbera,
-    Dmitry Bogdanov, Andrés Ferraro, Sergio Oramas, Alastair Porter, and
-    Xavier Serra. "Freesound Datasets: A Platform for the Creation of Open Audio
-    Datasets." International Society for Music Information Retrieval Conference
+    [2] Eduardo Fonseca, Jordi Pons, Xavier Favory, Frederic Font Corbera, 
+    Dmitry Bogdanov, Andrés Ferraro, Sergio Oramas, Alastair Porter, and 
+    Xavier Serra. "Freesound Datasets: A Platform for the Creation of Open Audio 
+    Datasets." International Society for Music Information Retrieval Conference 
     (ISMIR), pp. 486–493. Suzhou, China, 2017.
-
+    
     Args:
-        root (str): Folder where the FUSS data is. Either points to ssdata or
+        root (str): Folder where the FUSS data is. Either points to ssdata or 
           ssdata_reverb.
-        split (str): Either the ``train``, ``validation``, or ``eval`` split.
+        split (str): Either the ``train``, ``validation``, or ``eval`` split. 
         kwargs: Additional keyword arguments to BaseDataset.
     """
     def __init__(self, root, split='train', **kwargs):
@@ -458,9 +431,9 @@ class FUSS(Scaper):
             raise DataSetException(
                 f"split '{split}' not one of the accepted splits: "
                 f"'train', 'validation', 'eval'.")
-
+        
         folder = os.path.join(root, split)
-        super().__init__(folder, sample_rate=16000, strict_sample_rate=True,
+        super().__init__(folder, sample_rate=16000, strict_sample_rate=True, 
                          **kwargs)
 
     def _get_info_from_item(self, item):
@@ -485,9 +458,7 @@ class WHAM(MixSourceFolder):
     """
     Hook for the WHAM dataset. Essentially subclasses MixSourceFolder but with presets
     that are helpful for WHAM, which as the following directory structure:
-
     .. code-block:: none
-
         [wav8k, wav16k]/
           [min, max]/
             [tr, cv, tt]/
@@ -502,7 +473,6 @@ class WHAM(MixSourceFolder):
           cv/
           tt/
           metadata/
-
     Args:
         root (str): Root of WHAM directory.
         mix_folder (str): Which folder is the mix? Either 'mix_clean', 'mix_both', or
@@ -518,11 +488,11 @@ class WHAM(MixSourceFolder):
     }
 
     DATASET_HASHES = {
-        "wav8k": "acd49e0dae066e16040c983d71cc5a8adb903abff6e5cbb92b3785a1997b7547",
+        "wav8k": "acd49e0dae066e16040c983d71cc5a8adb903abff6e5cbb92b3785a1997b7547", 
         "wav16k": "5691d6a35382f2408a99594f21d820b58371b5ea061841db37d548c0b8d6ec7f"
     }
 
-    def __init__(self, root, mix_folder='mix_clean', mode='min', split='tr',
+    def __init__(self, root, mix_folder='mix_clean', mode='min', split='tr', 
                  sample_rate=8000, **kwargs):
         if mix_folder not in self.MIX_TO_SOURCE_MAP.keys():
             raise DataSetException(
@@ -647,15 +617,17 @@ class Slakh(BaseDataset):
             return len(sources) >= self.min_acceptable_sources
 
         trackpaths = [os.path.join(folder, f) for f in os.listdir(folder)
-            if num_instruments_acceptable(f)]
+                      if num_instruments_acceptable(f)]
 
         return trackpaths
 
 
     def submix(self, sources, num_frames):
         for source, values in sources.items():
-            sources[source] = (sum(values) if values else
-                 self._load_audio_from_array(np.zeros((1, num_frames))))
+            sources[source] = (
+                sum(values) if values else
+                self._load_audio_from_array(np.zeros((1, num_frames),
+                                                     dtype=np.float32)))
 
 
     def process_item(self, srcs_dir):
@@ -665,7 +637,7 @@ class Slakh(BaseDataset):
         src_metadata = yaml.load(open(os.path.join(srcs_dir, 'metadata.yaml'), 'r'), Loader=Loader)
         _, audio_dir = os.path.split(src_metadata["audio_dir"])
         _, midi_dir = os.path.split(src_metadata["midi_dir"])
-        
+
         sources = {}
         for source in self.sources:
             sources[source] = []
